revert(alert): rename cmds
