--- scripts/.k8s/update-alert.cronjob.yaml
@@ -1,38 +1,0 @@
-apiVersion: batch/v1beta1
-kind: CronJob
-metadata:
-  name: scripts
-  labels:
-    app.kubernetes.io/part-of: ${CI_PROJECT_NAME}
-    owner: ${CI_PROJECT_NAME}
-    team: ${CI_PROJECT_NAME}
-spec:
-  concurrencyPolicy: Forbid
-  successfulJobsHistoryLimit: 3
-  failedJobsHistoryLimit: 3
-  schedule: "*/15 * * * *"
-  jobTemplate:
-    spec:
-      backoffLimit: 0
-      template:
-        spec:
-          containers:
-            - name: update-alert
-              image: ${CI_REGISTRY_IMAGE}/scripts:${IMAGE_TAG}
-              resources:
-                requests:
-                  cpu: 1500m
-                  memory: 1.5Gi
-                limits:
-                  cpu: 2000m
-                  memory: 2Gi
-              workingDir: /app
-              env:
-                - name: PRODUCTION
-                  value: "${PRODUCTION}"
-                - name: GRAPHQL_ENDPOINT
-                  value: "http://hasura-cdtn-admin/v1/graphql"
-              envFrom:
-                - secretRef:
-                    name: cdtn-admin-secrets
-          restartPolicy: Never
