--- scripts/lib/ccn-list.js
@@ -1,55 +1,0 @@
-const ccns = [
-  { id: "KALICONT000005635624", num: 16 },
-  { id: "KALICONT000005635234", num: 29 },
-  { id: "KALICONT000005635613", num: 44 },
-  { id: "KALICONT000005635630", num: 86 },
-  { id: "KALICONT000005635184", num: 176 },
-  { id: "KALICONT000005635872", num: 275 },
-  { id: "KALICONT000005635856", num: 292 },
-  { id: "KALICONT000005635407", num: 413 },
-  { id: "KALICONT000005635373", num: 573 },
-  { id: "KALICONT000005635842", num: 650 },
-  { id: "KALICONT000005635617", num: 675 },
-  { id: "KALICONT000005635826", num: 787 },
-  { id: "KALICONT000005635886", num: 843 },
-  { id: "KALICONT000005635953", num: 1043 },
-  { id: "KALICONT000005635191", num: 1090 },
-  { id: "KALICONT000005635409", num: 1147 },
-  { id: "KALICONT000005635418", num: 1266 },
-  { id: "KALICONT000005635405", num: 1351 },
-  { id: "KALICONT000005635653", num: 1404 },
-  { id: "KALICONT000005635444", num: 1480 },
-  { id: "KALICONT000005635594", num: 1483 },
-  { id: "KALICONT000005635173", num: 1486 },
-  { id: "KALICONT000005635596", num: 1501 },
-  { id: "KALICONT000005635421", num: 1505 },
-  { id: "KALICONT000005635435", num: 1516 },
-  { id: "KALICONT000005635870", num: 1517 },
-  { id: "KALICONT000005635177", num: 1518 },
-  { id: "KALICONT000005635413", num: 1527 },
-  { id: "KALICONT000005635221", num: 1596 },
-  { id: "KALICONT000005635220", num: 1597 },
-  { id: "KALICONT000005635871", num: 1606 },
-  { id: "KALICONT000005635918", num: 1672 },
-  { id: "KALICONT000005635467", num: 1702 },
-  { id: "KALICONT000005635685", num: 1740 },
-  { id: "KALICONT000005635534", num: 1979 },
-  { id: "KALICONT000005635528", num: 1996 },
-  { id: "KALICONT000005635550", num: 2098 },
-  { id: "KALICONT000005635792", num: 2111 },
-  { id: "KALICONT000005635780", num: 2120 },
-  { id: "KALICONT000005635557", num: 2148 },
-  { id: "KALICONT000005635085", num: 2216 },
-  { id: "KALICONT000005635813", num: 2264 },
-  { id: "KALICONT000005635807", num: 2395 },
-  { id: "KALICONT000017941839", num: 2420 },
-  { id: "KALICONT000017577652", num: 2511 },
-  { id: "KALICONT000018563755", num: 2596 },
-  { id: "KALICONT000018773893", num: 2609 },
-  { id: "KALICONT000018926209", num: 2614 },
-  { id: "KALICONT000025805800", num: 2941 },
-  { id: "KALICONT000027172335", num: 3043 },
-  { id: "KALICONT000027084096", num: 3127 },
-];
-
-module.exports = { ccns };
