--- scripts/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "cdtn-admin-scripts",
-  "version": "1.0.0",
-  "dependencies": {
-    "@cdtn-admin/infra": "1.0.0",
-    "nodegit": "0.25.0",
-    "semver": "^7.3.2"
-  },
-  "devDependencies": {
-    "@rollup/plugin-commonjs": "^13.0.0",
-    "@rollup/plugin-node-resolve": "^8.0.1",
-    "builtin-modules": "^3.1.0",
-    "rollup": "^2.15.0"
-  },
-  "license": "MIT",
-  "main": "index.js",
-  "private": true,
-  "scripts": {
-    "build": "rollup -c",
-    "start": "node --enable-source-maps ./dist/update-alert.js",
-    "start:dev": "GRAPHQL_ENDPOINT=http://localhost:8080/v1/graphql HASURA_GRAPHQL_ADMIN_SECRET=admin1 yarn start",
-    "start:dump": "GRAPHQL_ENDPOINT=http://localhost:8080/v1/graphql HASURA_GRAPHQL_ADMIN_SECRET=admin1 DUMP=true yarn start"
-  }
-}
